requirements.txt
streamlit
psycopg2-binary
argon2-cffi